move tests for RetrySettings to its own class
